struct declaration in header, fix warning
