+++ Source/Diagnostics/ComputeDiagFunctors/BackTransformParticleFunctor.cpp
@@ -6,6 +6,55 @@
 #include <AMReX_Print.H>
 #include <AMReX_BaseFwd.H>
 
+SelectParticles::SelectParticles (const WarpXParIter& a_pti, TmpParticles& tmp_particle_data,
+                                  amrex::Real current_z_boost, amrex::Real old_z_boost,
+                                  int a_offset)
+    : m_current_z_boost(current_z_boost), m_old_z_boost(old_z_boost)
+{
+    if (tmp_particle_data.size() == 0) return;
+    m_get_position = GetParticlePosition(a_pti, a_offset);
+
+    const auto lev = a_pti.GetLevel();
+    const auto index = a_pti.GetPairIndex();
+
+    zpold = tmp_particle_data[lev][index][TmpIdx::zold].dataPtr();
+}
+
+
+LorentzTransformParticles::LorentzTransformParticles ( const WarpXParIter& a_pti,
+                                TmpParticles& tmp_particle_data,
+                                amrex::Real t_boost, amrex::Real dt,
+                                amrex::Real t_lab, int a_offset)
+    : m_t_boost(t_boost), m_dt(dt), m_t_lab(t_lab)
+{
+    using namespace amrex::literals;
+    
+    if (tmp_particle_data.size() == 0) return;
+    m_get_position = GetParticlePosition(a_pti, a_offset);
+    
+    auto& attribs = a_pti.GetAttribs();
+    m_wpnew = attribs[PIdx::w].dataPtr(); 
+    m_uxpnew = attribs[PIdx::ux].dataPtr(); 
+    m_uypnew = attribs[PIdx::uy].dataPtr();
+    m_uzpnew = attribs[PIdx::uz].dataPtr();
+
+    const auto lev = a_pti.GetLevel();
+    const auto index = a_pti.GetPairIndex();
+
+    m_xpold = tmp_particle_data[lev][index][TmpIdx::xold].dataPtr();
+    m_ypold = tmp_particle_data[lev][index][TmpIdx::yold].dataPtr();
+    m_zpold = tmp_particle_data[lev][index][TmpIdx::zold].dataPtr();
+    m_uxpold = tmp_particle_data[lev][index][TmpIdx::uxold].dataPtr();
+    m_uypold = tmp_particle_data[lev][index][TmpIdx::uyold].dataPtr();
+    m_uzpold = tmp_particle_data[lev][index][TmpIdx::uzold].dataPtr();
+
+    m_betaboost = WarpX::beta_boost;
+    m_gammaboost = WarpX::gamma_boost;
+    m_Phys_c = PhysConst::c;
+    m_inv_c2 = 1._rt/(m_Phys_c * m_Phys_c);
+    m_uzfrm = -m_gammaboost*m_betaboost*m_Phys_c;
+}
+
 /**
  * \brief Functor to compute Lorentz Transform and store the selected particles in existing
  * particle buffers
@@ -19,161 +68,6 @@ BackTransformParticleFunctor::BackTransformParticleFunctor (
     InitData();
 }
 
-struct SelectParticles
-{
-    using TmpParticles = WarpXParticleContainer::TmpParticles;
-
-    SelectParticles( const WarpXParIter& a_pti, TmpParticles& tmp_particle_data,
-                     amrex::Real current_z_boost, amrex::Real old_z_boost,
-                     int a_offset = 0)
-        : m_current_z_boost(current_z_boost), m_old_z_boost(old_z_boost)
-    {
-        if (tmp_particle_data.size() == 0) return;
-        m_get_position = GetParticlePosition(a_pti, a_offset);
- 
-        const auto lev = a_pti.GetLevel();
-        const auto index = a_pti.GetPairIndex();
-
-        zpold = tmp_particle_data[lev][index][TmpIdx::zold].dataPtr();
-    }
-
-    template <typename SrcData>
-    AMREX_GPU_HOST_DEVICE
-    int operator() (const SrcData& src, int i) const noexcept
-    {
-        using namespace amrex;
-        amrex::ParticleReal xp, yp, zp;
-        m_get_position(i, xp, yp, zp);
-        int Flag = 0; 
-        if ( ( (zp >= m_current_z_boost) && (zpold[i] <= m_old_z_boost) ) ||
-             ( (zp <= m_current_z_boost) && (zpold[i] >= m_old_z_boost) ))
-        {    Flag = 1;
-        }
-        return Flag;
-    }
-    
-    GetParticlePosition m_get_position;
-    amrex::Real m_current_z_boost;
-    amrex::Real m_old_z_boost;
-    amrex::ParticleReal* AMREX_RESTRICT zpold = nullptr;
-};
-
-
-struct LorentzTransformParticles
-{
-    
-    using TmpParticles = WarpXParticleContainer::TmpParticles;
-
-    LorentzTransformParticles ( const WarpXParIter& a_pti, TmpParticles& tmp_particle_data,
-                                amrex::Real t_boost, amrex::Real dt,
-                                amrex::Real t_lab, int a_offset = 0)
-        : m_t_boost(t_boost), m_dt(dt), m_t_lab(t_lab)
-    {
-        using namespace amrex::literals;
-        
-        if (tmp_particle_data.size() == 0) return;
-        m_get_position = GetParticlePosition(a_pti, a_offset);
-        
-        auto& attribs = a_pti.GetAttribs();
-        m_wpnew = attribs[PIdx::w].dataPtr(); 
-        m_uxpnew = attribs[PIdx::ux].dataPtr(); 
-        m_uypnew = attribs[PIdx::uy].dataPtr();
-        m_uzpnew = attribs[PIdx::uz].dataPtr();
-
-        const auto lev = a_pti.GetLevel();
-        const auto index = a_pti.GetPairIndex();
-
-        m_xpold = tmp_particle_data[lev][index][TmpIdx::xold].dataPtr();
-        m_ypold = tmp_particle_data[lev][index][TmpIdx::yold].dataPtr();
-        m_zpold = tmp_particle_data[lev][index][TmpIdx::zold].dataPtr();
-        m_uxpold = tmp_particle_data[lev][index][TmpIdx::uxold].dataPtr();
-        m_uypold = tmp_particle_data[lev][index][TmpIdx::uyold].dataPtr();
-        m_uzpold = tmp_particle_data[lev][index][TmpIdx::uzold].dataPtr();
-
-        m_betaboost = WarpX::beta_boost;
-        m_gammaboost = WarpX::gamma_boost;
-        m_Phys_c = PhysConst::c;
-        m_inv_c2 = 1._rt/(m_Phys_c * m_Phys_c);
-        m_uzfrm = -m_gammaboost*m_betaboost*m_Phys_c;
-    }
-
-    template <typename DstData, typename SrcData>
-    AMREX_GPU_HOST_DEVICE
-    void operator () (const DstData& dst, const SrcData& src, int i_src, int i_dst) const noexcept
-    {
-        using namespace amrex::literals;
-        // get current src position
-        amrex::ParticleReal xpnew, ypnew, zpnew;
-        m_get_position(i_src, xpnew, ypnew, zpnew);
-        const amrex::Real gamma_new_p = std::sqrt(1.0_rt + m_inv_c2*
-                                        ( m_uxpnew[i_src] * m_uxpnew[i_src]
-                                        + m_uypnew[i_src] * m_uypnew[i_src]
-                                        + m_uzpnew[i_src] * m_uzpnew[i_src]));
-        const amrex::Real gamma_old_p = std::sqrt(1.0_rt + m_inv_c2*
-                                        ( m_uxpold[i_src] * m_uxpold[i_src]
-                                        + m_uypold[i_src] * m_uypold[i_src]
-                                        + m_uzpold[i_src] * m_uzpold[i_src]));
-        const amrex::Real t_new_p = m_gammaboost * m_t_boost - m_uzfrm * zpnew * m_inv_c2;
-        const amrex::Real z_new_p = m_gammaboost* ( zpnew + m_betaboost * m_Phys_c * m_t_boost);
-        const amrex::Real uz_new_p = m_gammaboost * m_uzpnew[i_src] - gamma_new_p * m_uzfrm;
-        const amrex::Real t_old_p = m_gammaboost * (m_t_boost - m_dt)
-                                    - m_uzfrm * m_zpold[i_src] * m_inv_c2;
-        const amrex::Real z_old_p = m_gammaboost * ( m_zpold[i_src] + m_betaboost
-                                                     * m_Phys_c * (m_t_boost - m_dt ) );
-        const amrex::Real uz_old_p = m_gammaboost * m_uzpold[i_src] - gamma_old_p * m_uzfrm;
-        // interpolate in time to t_lab
-        const amrex::Real weight_old = (t_new_p - m_t_lab)
-                                     / (t_new_p - t_old_p);
-        const amrex::Real weight_new = (m_t_lab - t_old_p)
-                                     / (t_new_p - t_old_p);
-        // weighted sum of old and new values
-        const amrex::ParticleReal xp = m_xpold[i_src] * weight_old + xpnew * weight_new;
-        const amrex::ParticleReal yp = m_ypold[i_src] * weight_old + ypnew * weight_new;
-        const amrex::ParticleReal zp = z_old_p * weight_old + z_new_p * weight_new;
-        const amrex::ParticleReal uxp = m_uxpold[i_src] * weight_old
-                                      + m_uxpnew[i_src] * weight_new;
-        const amrex::ParticleReal uyp = m_uypold[i_src] * weight_old
-                                      + m_uypnew[i_src] * weight_new;
-        const amrex::ParticleReal uzp = uz_old_p * weight_old
-                                      + uz_new_p * weight_new;
-        dst.m_aos[i_dst].pos(0) = xp;
-#if (AMREX_SPACEDIM == 3)
-        dst.m_aos[i_dst].pos(1) = yp;
-        dst.m_aos[i_dst].pos(2) = zp;
-#elif (AMREX_SPACEDIM == 2)
-        dst.m_aos[i_dst].pos(1) = zp;
-        amrex::ignore_unused(yp);
-#endif
-        dst.m_rdata[PIdx::w][i_dst] = m_wpnew[i_src];
-        dst.m_rdata[PIdx::ux][i_dst] = uxp;
-        dst.m_rdata[PIdx::uy][i_dst] = uyp;
-        dst.m_rdata[PIdx::uz][i_dst] = uzp;
-    }
-
-    GetParticlePosition m_get_position;
-    
-    amrex::ParticleReal* AMREX_RESTRICT m_xpold = nullptr;
-    amrex::ParticleReal* AMREX_RESTRICT m_ypold = nullptr;
-    amrex::ParticleReal* AMREX_RESTRICT m_zpold = nullptr;
-
-    amrex::ParticleReal* AMREX_RESTRICT m_uxpold = nullptr;
-    amrex::ParticleReal* AMREX_RESTRICT m_uypold = nullptr;
-    amrex::ParticleReal* AMREX_RESTRICT m_uzpold = nullptr;
-
-    const amrex::ParticleReal* AMREX_RESTRICT m_uxpnew = nullptr;
-    const amrex::ParticleReal* AMREX_RESTRICT m_uypnew = nullptr;
-    const amrex::ParticleReal* AMREX_RESTRICT m_uzpnew = nullptr;
-    const amrex::ParticleReal* AMREX_RESTRICT m_wpnew = nullptr;
-
-    amrex::Real m_gammaboost;
-    amrex::Real m_betaboost;
-    amrex::Real m_Phys_c;
-    amrex::Real m_uzfrm;
-    amrex::Real m_inv_c2;
-    amrex::Real m_t_boost;
-    amrex::Real m_dt;
-    amrex::Real m_t_lab;
-};
 
 void
 BackTransformParticleFunctor::operator () (ParticleContainer& pc_dst, int &totalParticleCounter, int i_buffer) const
@@ -190,7 +84,6 @@ BackTransformParticleFunctor::operator () (ParticleContainer& pc_dst, int &total
         amrex::Real dt = warpx.getdt(0);
 
         for (WarpXParIter pti(*m_pc_src, lev); pti.isValid(); ++pti) {
-            auto index = std::make_pair(pti.index(), pti.LocalTileIndex());
             auto ptile_dst = pc_dst.DefineAndReturnParticleTile(lev, pti.index(), pti.LocalTileIndex() );
         }
 
@@ -208,7 +101,6 @@ BackTransformParticleFunctor::operator () (ParticleContainer& pc_dst, int &total
 
                 auto index = std::make_pair(pti.index(), pti.LocalTileIndex());
 
-                const auto GetPosition = GetParticlePosition(pti);
                 const auto GetParticleFilter = SelectParticles(pti, tmp_particle_data,
                                                m_current_z_boost[i_buffer],
                                                m_old_z_boost[i_buffer]);
